Establish rendering pattern for SVG elements
This is basically what we were doing within the Rails app just with a
little more extension. It’s a bummer but React doesn’t support the `use`
and `def` tags fully. There are some other libraries out there, just not
sure it’s something we need to get into.