Dump redux-react-router until it’s stable.
It’s a little bumpy with all the warnings it’s generating from React and 
there really isn’t any movement on the package itself. It seems like the
“rackt” group may provide their own wrapper around react-router or it 
will be solved in that project specifically.

I’m watching https://github.com/acdlite/redux-react-router/issues/7 for
any resolution and will potentially bring it back in if that ever gets
fixed.