Bring back PropTypes and some basic type checking
It’s actually pretty handy in development as it means we are coding 
towards interfaces rather than _hoping_ or constantly checking for the
property/method to exist. The propTypes blocks are pretty handy for a
quick glance to see what is actually getting injected or what the
expectation of what gets injected is.

Also we removed another override in the linter. More work to happen here
as we dive deeper for sure.