Add modals which acts as a container for alerts
Probably some more work to do here when we get a little more serious
around modals. We can apply the basic concept of modals to alerts for
the time being. There’s problems with this logic (multiple alerts) but
this is good ground work and should cover most of the needs we have for
on-boarding.