Adds the follow all for the community picker
I also abstracted the pickers up a level since they are sharing a bit
of the same behavior. There are more opportunities here to pull things 
apart (CSS, render, actions) but I’d like to wait till we see a picker 
outside of on-boarding first to see what that looks like. 

/cc @rynbyjn