Fine tuning the scroll behavior (see notes below)
Not much really happened here in the final commit but I did a lot of
exploring with the navbar sliding out of the way. The main experiment
was using `translateY` based on the percentage of the distance scrolled
vs the height of the navbar. It’s rad because you can stop it halfway
and see a good part of the navbar, but it’s really clunky when zipping
up or down the page. The CSS transitions just offer a more performant 
way to handle this. It was a good experiment and I think where we ended
up is a good place.