Adds the banners to logged out search.
Sort of a bummer here but we broke from the extension. The binding’s 
get super confused on the calls out to super and throw errors all over
the place. We may change this down the road, but to keep this moving
going to just decouple and duplicate the logic for now.