Create relationship layouts in a post header
I made a few adjustments around reposts and tried to normalize down
some of the craziness, so we’ll see if she flies.

There’s a little bit of magic and some cleanup to be done here but it’s 
in way better shape. Thinking about waiting till we get some feedback
first. :boom: