Create a RelationsGroup to manage relationships
Non-connected pure functions need to be able and create relationship
components based on the layouts. We also need to DRY up the callbacks
for syncing the relationship changes back to the store. This is a start
to all of that.

The goal is to eventually make RelationsGroup a dumb component and have
containers inject the handlers. This at least is allowing us to move 
the actions and handlers into a single spot. 

It definitely mucks up the UserCard but we’ll clean that up in a
future PR after logged out re-launch.

And of course on-boarding and the batch saving within Pickers makes 
things a bit more complicated. Also something to rethink down the road.