Set Text/ImageTegions to be clickable in grid mode
The image is easy, text is a little trickier since it can have nested
links inside of it. The TextRegion has a listener setup to capture any
clicks within the RegionContent element and then trigger a click on a 
hidden “More” button. This has a nice side effect in that we have a link
that still describes what it’s going to do. The downside to this is it
looks like it breaks opening in a new tab. Something to potentially 
revisit down the road.