Temporarily rollback background images for Avatars
The problem with this and ultimately what is the blurry image was that
the default regular image size was only at 60px when it needed to be at
120px. You can see that change at ello/ello@8a1974c. That deployed this
morning, but it’ll take about 24 hours for it to propagate through.

For the meantime, we rolled back to the background-image version which
fills the circle but still has a blurry image to it. Again, once that
propagates through the image should be at a much better resolution.