Connect TextRegions to handle relative links
This allows links that are set as relative: usernames, markdown links, 
raw links and eventually hashtags to be sent through the router. We
still have the special hashtag case in here but that can eventually get
removed once we change it in the API.

Eventually we’ll need to deal with rerouting all hashtag links through
`/find` but that can wait till we release the web app to logged in users.

Wired up the linking logic on reposted items as well.