Add loader states to ImageRegions and fade up
I didn’t add the loading animation here as I’m not sure if it’s needed
or not? It’d be just adding the styles/animations at this point as the
states are already getting set. From a performance standpoint it’d be 
better not to have one though.