Change the flow for uploading profile assets
I added a `tmp` key on the two image assets (avatar, coverImage) while 
an image is uploading. This gives us the immediate feedback and when
the image is done, we need clear out the tmp variable so we’re not 
throwing it into storage. It doesn’t work totally as expected and gets
a little weird but I’m sure it’s something minor.

We’ll need to ensure when a user does upload something, that we merge 
their profile from within `json.users` so they propagate through.

This also moves all of the temporary asset handling up to each asset
instead of their containers. This way the container only cares about
passing them the key (avatar, coverImage) out of the profile and the
asset can take care of which one to render based on size and existence.

This has on-boarding changes in it, but hasn’t been test yet. We’ll
break some more stuff first.