Fixes the layout for the button in a flag dialog
The top value is determined when the dialog is launched and rendered 
which means the button doesn’t respond to resize events within the 
browser. This shouldn’t affect things like orientation change since 
updates are between much bigger breakpoints (mobile, tablet, desktop).

I’d prefer not to add this in, as it seems like overkill to have a 
listener on this thing to cover a pretty small use case.

[Finishes: #111009942]