Tune mobile completers and allow them to scroll
We are not able to “attach” the completers to the soft keyboard like the
iOS app as we don’t have access to it’s dimension at the browser level. 
This is sort of a bummer. What I did for mobile was keep it just below 
the text being completed and give it a max height of 153 points and 
allow it to scroll through the list. I’ve only tested this in the 
simulator so far. I’m sure it’ll take a couple of rounds to get it 
right, it’s definitely not perfect.

Also set a max height for the desktop completer at 320 points and let it
scroll.

@rynbyjn, we might be cool returning 20+ results here unless the API 
crawls?