Move the APP_VERSION into app instead of config.
It felt kinda silly to have to remember to update a config var before pushing up a deploy. This allows us to just modify it in the js and have it purge from one spot. Also allows us the ability to have more logic around what gets purged if we want to be more specific about it in the future, but it purges all for now. this probably won’t need to be updated as much now that we aren’t persisting much data. Would only need to update if we changed the format of authentication, editor, gui, json(deleted arrays for posts/comments/users), or an update to the profile json. Once this gets merged in we can remove the config var APP_VERSION from heroku.