Allows appending content to TextBlocks
This allows the TextBlock to re-render once it’s data value no longer is the same as it’s inner markup. This means it’s outside influence can append content or completely rewrite it. This is a good thing.

The guard in `shouldComponentUpdate` still prohibits the re-render of the element as the User is typing since those values should always be equal.

Appending text content alway happens in the very end of the very last block.

Also cleaned up some formatting of some upcoming linter changes.