Set the scroll position for templates with a Cover
Another try getting this to play well with the React rendering cycles. The Navbar handles the scroll for blacklisted templates (Settings and UserDetails) and calls the `scrollTo` method only when the route changes. The individual View controllers handle calling the `scrollTo` when they are mounted. This ensure the readiness of the page and various objects. This also specifically sets the state of the navbar `asLocked` in these templates.

:spaghetti:


[Finishes: #108011058]
[Fixes: #112849785]
[Fixes: #113849111]