Rebuild the inner structure of notifications
We’ve now separated the NotificationAsset (ImageRegion/EmbedRegion) and combined the NotificationBody (notification message) and NotificationSummary text blocks (TextRegions) into a single element. This allows us to create 2 columns when there is an asset that splits the Notification. This means the asset is no longer positioned absolutely and has an actual height and won’t create overlap. When there is no asset we now can now let the NotificationBody span the entire width which looks much better.

Having this setup means we can truncate the username and wrap long links within a NotificationBody.

[Finishes: #115289863]
[Finishes: #104755086]
[Finishes: #115289753]