Cleans up the loading state of the Avatars
This might not be perfect yet but it should be a little less jarring going between pages. We still have a loading state but without the pulsating Avatars. The alternative here is to delay the creation of the loader and swapping classes when a load is needed which could get a little tricky here.

[Finishes #115638987]