Clears out the BatchEmailControls text on submit
[Finishes: #115642829]

I tried to fake the response we’d need to immediately render the Avatar in the list but this proved troublesome and resulted in duplicate users ending up in that list. There’s some crucial information we need from the server response first. If we really want this, we should write a few API stories around it. Not sure it’s the highest priority right now?