Rolls back react-helmet
The 3.0.0 version of [react-helmet](https://github.com/nfl/react-helmet/) breaks the class attribute on the html element. This is where we store things for CSS feature detection so it broke a bunch of functionality and styling for hover states. I’m subscribed to the [open issue](https://github.com/nfl/react-helmet/issues/121) and will keep an eye on things.

[Finishes: #116486511]
[Finishes: #116487199]