Hide Credits that potentially cover a FormControl
This is only for Android and was occurring in multiple views.

It seems some versions of Android calculate the height of the viewport as the remaining space when the keyboard is open so 100vh is really tiny. Since Credits are positioned bottom this moves them above inputs and all other stuff. iOS seems to treat their soft keyboard as modal like so 100vh is the viewport height.

[Finishes: #116472155]
[Finishes: #116471825]
[Finishes: #116472431]