Fix a weird repainting issue with the more tool
There is something Safari and Mobile Safari does not like about the positioning of the More button. It wouldn’t appear unless a repaint was triggered which could be as simple as just hovering over where the icon is or the hot reloading of the styles in development. It turned out to be the rotation on the `SVGIcon > g` element. I just moved that rotation up to the `SVGIcon` and it seems to have fixed this issue in the simulator and desktop safari at really small screens. :shrugs:

[Finishes: #116554221]