Start tracking layouts with an offset in the gui
Enough is enough with these guys. These funky offsets occur in Settings and the UserDetail pages. This just starts to move the means of finding them out and into the gui reducer. I’m going to start storing more info around the current gui state in this guy. We could benefit from storing a `template` name too but I’m not entirely sure we need to yet. This just gives things like scroll position and eventually a CSS modifier down the road to target these layouts a little easier. 

Those TODO’s are valid, it’s just duplicated code at the moment and will get cleaned up in the next couple of PRs as we move through these files.

This also removes username and pathname from the FooterContainer. It’s still calling window.scrollTo here, but again that’ll get cleaned up with the above most likely while we are tackling the Navbar.