WIP: Preliminary work breaking apart the Navbar
I’ll add more details and a full PR description once this is worked through a little more but for now…

- Breaking the logic apart into a Container/Presentational component
- Converting the logged out Navbar to a stateless function
- Starting to break out the gui reducer to handle things like `isGridMode`, `isNavbarFixed`, `viewportDeviceSize`, etc. This should encompass more of these global properties that we are storing from the `ResizeComponent` and the `ScrollComponent` among others.
- Hook into the gui reducer instead of the global properties
- Introduce the `ViewportContainer` and `Viewport` for toggling classes based on the current state of the `gui` reducer and other gui related items.
- Limit the number of renders the Navbar and Footer encounter.
- Update styles to handle states from the `Viewport
- Export pure components individually without their defaults (new import types)

The idea behind the `ViewportContainer` is to describe the state of the application in a more generic sense. While it make sense to normally do this directly on say the `Footer` or the `Navbar` it was generating a lot of needless render calls down to the components. While not a terrible thing, it’s definitely not the most performant. By offloading it to the `ViewportContainer` and thus `Viewport` which will never have any children, the render only hits one node and it’s really just the `classList` that  updates. In my early testing I’ve seen some big wins here from not calling re-renders all over the place. Especially with scrolling. 

:point_up: Hopefully that’s enough of a description to remind me what the hell I’m doing.