WIP Continue sticking more in the gui reducer
Still breaking apart the containers and components for Navbar, Viewport and a little bit in the Footer. Most of the work continues to be moving logic up into the gui reducer and allowing Containers to subscribe to those property changes. This group of changes puts the engine back together minus the key commands.