Don’t require gui reducer props at a top level
My theory is this is a race condition and the gui reducer hasn’t been created yet. I guess this is possible since both NavbarContainer and ViewportContainer are created in App and ViewportContainer is the first one to read or write to that reducer. But it doesn’t fully make sense since the values it was failing on `isGridMode`, `isProfileMenuActive` and `isNotificationsActive` are actually initialized with values… ? And they are required in their Views Navbar/Viewport and those don’t throw errors.

Yay! 

Regardless this fixes the issue which we were seeing running `localStorage.clear()`.