Scroll offset layouts for server rendered pages
Attempting to fix the jumpiness that only happens when isomorphic rendered pages are involved in Settings or UserDetail. Normally we call the `scrollToOffsetTop` in `componentDidMount` which works great for client rendered pages but doesn’t fire for ones coming from the server. This means `window.scrollY` remains at `0` until the client has had a chance to download, initialize and run `componentDidMount`. 

We weren’t calling it before for the server so my hope is that this fixes it “enough”. It’s still being called after React is rendered so it could potentially still not be early enough. It kind of needs to happen here otherwise the `document.body > #root` doesn’t have any sort of height or even contents to it. I’m pretty sure this about as early as the mothership could call it too. We don’t have any crazy logic in there for this.

fingers crossed.

If this doesn’t work we could potentially…

Try setting the pages opacity at zero until after we’re able to call the `scrollTo`. My concern here would be SEO and how bots perceive it. Plus we would have to add / remove classes right at startup and it might feel a little laggy.

Another option is to start treating Cover images similar to the Promo images we use on Discover/Search. They would then be part of the Container instead of being a background image. We could make them clickable to expand them full screen instead? This option would make these layouts much, much simpler from a code perspective but definitely take a little reworking.


[Finishes: #118117741]