Move TextTools/Completer out to EditorContainer
- [x] Pull out of AppContainer so we can deal with them separately.
- [x] Move EditorTool states to modal reducer and into the redux flow
- [x] Reduce the amount of calls making its way to EditorTools render
- [x] Reduce the props AppContainer was watching

This isn’t perfect but it’s a good start to separating the concerns. There is definitely some work to do later within the EdtiorContainer and it’s subsequent components but this should be good for now.

Also the modal reducer is a bit of a mess, but I think this is something that gets reworked alongside gui. The modal reducer is basically gui without saving it to local storage. It will eventually have the omnibar reducer folded into it with a possible name change. It sort of depends what ends up in it? Right now it’s just states and properties for modal like objects. Kind of acting like a junk drawer.

I think once we’ve teased more of the gui apart and roped it into the redux flow we’ll have a pretty good idea of what can stay in gui/modal or switch between the two.