Remove GUI viewport constant from ResizeComponent
This gets rid of the global variable and allows the objects that were using this to pull these from the gui reducer instead. In most cases we just needed to pass the properties through (CommentTools/PostTools > FlagDialog) and in others we ended up connecting the component (ImageRegion). We’ll eventually split ImageRegion into a Container and a Component but for now we are just trying to clean up things to utilize the gui reducer instead.