Move ResizeComponent listeners to use gui reducer
Since we are now storing all of these properties on the gui reducer there really isn’t a need to have individual components listening to resize events. Instead we inject these presentational components with the correct properties from there Container components.

The only weird one here was StreamComponent as we just set it’s state with the resize components properties. I made it a little more specific for what it was using. Also removed some of those `window.innerHeight` to utilize what was set from the gui reducer.