Remove a local cache on the isAndroid lookup
We are handling the cache at the helper method now. 

I really wanted to move this into the reducer but it’s just not worth it. It’s my daily gift to the Android Gods that they get their shit together and then we don’t have to do stupid crap like this.