Roll back the findBy lookup to last stable version
This was introduced with https://github.com/ello/webapp/pull/619. What was happening was references were getting lost when we converted the for loops to use the `Object.keys`. This meant on your own UserDetail you wouldn’t see anything and on other people’s UserDetail you would get a random profile and header but the correct stream content. Not sure why but it has to do with being able to break out of that second loop. We’ll eventually want to fix this, but getting this fixed up in master is a higher priority for now.

/cc @rynbyjn

[Fixes: #119664781](https://www.pivotaltracker.com/story/show/119664781)