Ensure the onResize handler is fired at least once
We were guarding too much across sessions on updating the resize attributes. This _should_ fix the bug where Users were seeing only 2 columns until a resize event was triggered. 

This loosens the guard up some to call it at least the first time.

[Fixes: #119945085](https://www.pivotaltracker.com/story/show/119945085)

I also just updated a method name… we used to set these attributes on a constant, but we’re just returning them to be mapped to the store.