Adds optional action type and name check to specs
This is more of a linting type check than anything. What it’s trying check is having the action.type (constant) be similar to the action name (fn.name). So `GUI.SET_IS_OFFSET_LAYOUT` lower cased and camelized minus the GUI is equal to the action `setIsOffsetLayout`. It makes it much easier to match the types to the constants when trying to reason about the application. Or more accurately add, search, replace and remove actions and constants from your editor. 

It doesn’t translate well for generic actions like `LOAD_STREAM` but does work for a good chunk of the others.