Fixes an issue with last-of-type vs last-child
I’m not really sure why I was using `last-of-type` here but it was causing a weird bug when you had a `<ul> ~ <p> ~ <ul>` where the `<p>` element had a margin-bottom of `0`. This fixes that issue and doesn’t seem to cause any other visual regressions that I can make out.

[Finishes: #119114651](https://www.pivotaltracker.com/story/show/119114651)