Breaks apart PostDetail to Container/Component
The only real concern that needs to be tested is the removal of the `key` attribute on some of the nested elements. We shouldn’t need them since we are invalidating on the parent (PostDetailContainer), but that stuff is still kinda weird to me. If so, we might have to turn it into a class since pure components can’t have keys.

Took the render count down from 8 to 2 which given the work the meta tag generation is doing before hitting the virtual DOM should be a nice improvement.

[#120139929](https://www.pivotaltracker.com/story/show/120139929)