Allow call to action links to be passed to promos
Allows the author of the promotions to add a call to action button similar to the “Sign up” button that is the default for logged out users. 

To use the call to action button, it just needs to be specified as a normal html link within a promotion.

```
cta: <a href="https://ello.co/wtf" target="_blank">Learn more</a>,
```

This should give the author complete control of the link, it’s behavior and it’s display text.

Here are some rules.

1. For logged out users, including the call to action will take precedence over the default link (Sign up).
2. For logged in users, not including it will not show any link
3. We test that it’s an `<a>` element
4. We test for existence and simple validity for the href element
5. We test for `_blank` if there is a target.

[Finishes: #118439057](https://www.pivotaltracker.com/story/show/118439057)