Add a GooglePlayStore link to Authentication Views
I did a bit of refactoring here to use an image sprite instead of svg
assets. The GooglePlayStore link is pretty big to include on the page
as an svg. The compressed 2x sprite (AppleStore/GooglePlayStore) as an
image is about half the size of a compressed svg asset.

I also combined the AppleStore and GooglePlayStore under
assets/AppStores and removed the old support/AppleStoreLink

[Finishes: #125324301](https://www.pivotaltracker.com/story/show/125324301)