Add dimensions to basic ImageRegion attachments
While a post asset is still processing on the server we get a basic
attachment back. This means we don't have any preset image dimensions to
render with and images were not the correct scale. This fixes that bug
and specifically sets the image dimensions on basic attachments after
the image has been loaded into memory and we can retrieve it's size.
This should only affect that very short time frame when we have images
in this state. This occured while editing or potentially in post
creation in list view.

[Finishes: #126642709](https://www.pivotaltracker.com/story/show/126642709)