Optimize and add more unit tests around Discover
This breaks the page title out of the category selector so we can
calculate that on each page change rather than running the complicated
selector logic. It also tunes the shouldContainerUpdate method and tests
the exported functions. There's some more work to be done here as far as
breaking out more selectors but this would be better tackled down the
line. We wiouldn't really get a whole of performance benefits breaking
it out now. The shouldContainerUpdate should do the bulk fo the work
here.

[#126682613](https://www.pivotaltracker.com/story/show/126682613)