Add user.name and h1 to a UserDetail profile name
This changes the name from user.username to user.name if it exists and
replaces the h2 tag with an h1 for the main UserDetail profile. This
also reverts some of the avatar changes from the previous commit. We are
now just passing in whether it's the main UserDetail profile or another
profile.

[#126717975](https://www.pivotaltracker.com/story/show/126717975)