Update existing meta tag titles
This moves meta tags that never change up to the html template files.
This keeps the potential changes helmet has to look at a little more
focused. I also updated a few and removed a few that are no longer in
use. Most of this is in reference to https://github.com/joshbuchea/HEAD.

[Finishes: #126717975](https://www.pivotaltracker.com/story/show/126717975)