Delay navigating away on logout and delete account
We're going to wait an extra tick of the frame after clearing local
storage. This potentially could fix the issue where users are getting
stuck in infinite loops There may be more to do here, but the idea is to
get this out and see if it fixes the issue.

[Fixes: #130161029](https://www.pivotaltracker.com/story/show/130161029)
[Fixes: #130043483](https://www.pivotaltracker.com/story/show/130043483)