Break apart the HeroContainer and test selectors
Everything made it to it's home for components, constants and some
selectors. The selectors that are specific to the HeroContainer I left
as is until they need access somewhere else. One of the additional
routing selectors can probably be spread across other parts of the app.
It's really just a centralized way to map pathnames that seem to be
related to what view we're rendering.