Refactor and add Auth routes to Heros
Pretty big set of changes here allowing Hero to manage more of it's
contents. Breaks Hero out to HeroContainer, HeroRenderables and
HeroParts. Similar to how User is handled. This uses BackgroundImage to
replace Covers and bring Promotions, Credits directly into the
HeroContainer. It moves the Promotions and Covers out of Authentication
layouts. A few other cleanups around the asset components.