Use styles to highlight UserStats active states
This might not be permanent but this update removes the pathname key
passed down to the UserStats tab Links. The key was needed to force the
tablist to swap the active state. The bummer here is everything in the
UserProfile updates including the Avatar. If we use a little (gnarly)
CSS we can update the state without actually redrawing it in the DOM.
Super helpful not having the DOM update while we are scrolling the page.