Adjust layouts for the CTA and credits on mobile
This allows the CTA to expand height wise and truncates the text of the
username when the width is greater than 60%. The max width of the CTA is
40%.

This introduces a couple of other restrictions. The caption text cannot
be vertically centered within the promotion and occasionally the avatar
may get cutoff depending on the amount of text in the credits area and
the width of the device. The other restriction here is the credits cell
is always 60% the width regardless if there is a CTA or not.

Seems like a regression to the desktop layouts but whatever.

[Finishes: #134086341](https://www.pivotaltracker.com/story/show/134086341)