Add a dismiss icon to the bio modal
This allows an out for smaller devices like the iPhone SE.

This only adds it to the bio modal and not the remaining modals which
we've talked about. I'd like to leave that task to when we break dialogs
out to DialogRenderables. Something we can do in the future off of
master instead.

[Finishes: #137124933](https://www.pivotaltracker.com/story/show/137124933)