Moves global css back to CSS
We were loading global CSS through glamor, but the insert of it is
failing on the client. We were able to get it to work, but it would
require mucking up the server files. At the end of this thing, there
will be very little global CSS so we can just embed it into a style tag
within the head. For now, this commit just adds those global styles back
into the base.css file.

[Finishes: #141335023](https://www.pivotaltracker.com/story/show/141335023)